ipcache: move IPCache providers to ipcache cell
After removing the `newPolicyTrifecta` function with #34332, the
`IPCache` is still provided by functions in the daemon package.

This commit moves the function into its own cell in the package
`pkg/ipcache/cell`. The reason for the subspackage is to break
circular build dependencies.
